==> jupyter_lab_notebooks/3 - UMAP_cluster_projection.ipynb <==
base_path <- "/home/jovyan/storage/data"

# get the cell meta-data
path <- paste0(base_path, "/seurat_cell_meta_data.csv")
cell_meta_data <- read.csv(file = path, row.names = 1)

# load the cell-to-cell mapping matrix
path <- paste0(base_path, "/cell-to-cell_mapping.csv")
gw <- read.csv2(path, sep = ",", dec = ".", stringsAsFactors = FALSE, header = T, row.names = 1)
rownames(gw) <- as.character(sapply(X = rownames(gw), FUN = function(x){strsplit(x = x, ".", fixed=T)[[1]][1]}))

# load the
path <- paste0(base_path, "/confocal_states0-FilterWUSCLVtop100.csv")
ref <- read.csv2(path, row.names = 1, sep = ",")

# normalize locations (confocal cells)
gw1 <- gw/colSums(gw)  # make columns (physical cells) sum to 1

# create named vector mapping cells to clusters for scRNA-seq DGE
meta <- as.character(cell_meta_data$cluster_annotation)
names(meta) <- substr(cell_meta_data$cell_id, start = 1, stop = 16)

# obtain cluster numbers for cells in gw matrix
MATCH <- match(rownames(gw), names(meta))
cluster_numbers <- as.character(meta[MATCH])

# summed probabilities over all cells in 3D meristem for all clusters
prob_cluster_df <- NULL
for(cluster in unique(cluster_numbers)){
  gw_tmp <- gw1[cluster_numbers %in% cluster,]
  prob_cluster_df <- cbind(prob_cluster_df, colSums(gw_tmp))
}
colnames(prob_cluster_df) <- unique(cluster_numbers)

# add coordinates of cells
prob_cluster_df <- data.frame(ref[,1:3], prob_cluster_df)

prob_cluster_df[1:4,1:5]

# save the generated cluster projections
path <- paste0(base_path, "/cluster_assignment_3d.csv")
write.csv(x = prob_cluster_df, file = path)
