==> jupyter_lab_notebooks/5 - Evaluate Prediction Performance.ipynb <==
### 1 - Read in data
We need the expression values for all cells of all genes in the scRNA-seq dataset. We also need to know which genes have been used in the spatial reference expression dataset. In the following we read in those two datasets.

base_dir <- "/home/jovyan/storage"

# path to scRNA-seq DGE
dge_path <- paste0(base_dir, "/data/dge_day4_Oct-Nofilter_downsize.csv")

# path to spatial reference expression matrix
coord_path <- paste0(base_dir, "/data/confocal_states0-FilterWUSCLVtop100.csv")

# path to predict 3D gene expression profiles
sdge_path <- paste0(base_dir, "/data/sdge.csv")

# Load scRNA-seq dataset
dge <- read.csv(file = dge_path, row.names=1)
dge <- as.data.frame(t(dge))

# Load spatial expression reference dataset
ref <- as.matrix(read.csv(file = coord_path, row.names = 1, sep=","))

# Load predicted 3D gene expression profiles
sdge <- read.csv(file = sdge_path, row.names = 1)

### 2 - Calculate PEP-scores for genes in scRNA-seq datatset

First we obtain a correlation matrix that tells us how much each gene in the scRNA-seq dataset correlates with the genes in the spatial expression reference dataset. The resulting matrix has the dimension (scRNA-seq genes x reference genes).

genes <- intersect(colnames(dge), colnames(ref))
co <- cor(x = dge, y = dge[, genes], method = 'spearman')
dim(co)

Next we find the reference gene that has the highest correlation for all genes in the scRNA-seq dataset and record the correlation value. This value represents the PEP-score.  
The resulting data.frame contains the PEP-score for each gene in the scRNA-seq dataset as well as the reference gene that the correlation is based on.

# get top1 correlation for all genes with ref. genes
cor_top1 <- NULL
for(gen in rownames(co)){
  if(gen %in% colnames(co)){
      ord <- order(co[gen,], decreasing = TRUE)[2]
  } else {
      ord <- order(co[gen,], decreasing = TRUE)[1]
  }
  row <- c(as.numeric(co[gen, ord]), colnames(co)[ord])
  cor_top1 <- rbind(cor_top1, row)
}
cor_top1_df <- as.data.frame(cor_top1)

rownames(cor_top1_df) <- rownames(co); colnames(cor_top1_df) <- c('PEP-score', 'top_cor_gene')
cor_top1_df$`PEP-score` <- round(as.numeric(cor_top1_df$`PEP-score`),2)
cor_top1_df[1:4,]

### 3 - Interpretation of the PEP-score
The PEP-score can be used as an approximation on how confident we are that the model we trained in the notebook "1 - Predict_3D_gene_expression_profiles.ipynb" can give us (biologically) reasonable results. Since the score is only based on the expression of genes in the scRNA-seq and not the parameters that we used for the predictions, the PEP-score can only be used as a rough approximation for the out-of-sample prediction performance.  
The higher the PEP-score for a de novo 3D gene expression prediction, the more confident we are that the prediction makes sense biologically.

### 4 - Connection to the AUCROC
The prediction performance for..

install.packages("pROC")

library(pROC)

calculate_AUC <- function(type="roc", return_type = "auc", pre, ref){
      defaultW <- getOption("warn") 
      options(warn = -1) 

      auc_list <- NULL
      genes_tair <- intersect(colnames(ref), colnames(pre))
      for(gene_tair in genes_tair){
        if(type == "roc"){
            suppressMessages(
              aucroc <- roc(as.numeric(ref[, gene_tair]), as.numeric(pre[, gene_tair]), plot=FALSE, legacy.axes=TRUE, direction = "<")
            )
          auc_list[[gene_tair]] <- aucroc$auc[1]
        }else if(type == "pr"){
            suppressMessages(
              aucpr <- cvAUC::AUC(predictions=as.numeric(pre[, gene_tair]), labels=as.numeric(ref[, gene_tair]))
            )
          auc_list[[gene_symbol]] <- aucpr
        }
      }
    options(warn = defaultW)
    return(auc_list)
}

sdge[1:4,1:4]

auc <- calculate_AUC(pre = sdge, ref = ref)
auc <- data.frame(auc)
auc <- as.data.frame(t(auc))

auc
